+++ app/frontend/src/javascript/components/plans/plans-list.tsx
@@ -42,8 +42,10 @@ const PlansList: React.FC<PlansListProps> = ({ onError, onPlanSelection, onLogin
   const [groups, setGroups] = useState<Array<Group>>(null);
   // currently selected plan
   const [selectedPlan, setSelectedPlan] = useState<Plan>(null);
-  // filter shown plans by only one group
+  // filtering shown plans by only one group
   const [groupFilter, setGroupFilter] = useState<number>(null);
+  // filtering shown plans by ids
+  const [plansFilter, setPlansFilter] = useState<Array<number>>(null);
 
   // fetch data on component mounted
   useEffect(() => {
@@ -160,10 +162,27 @@ const PlansList: React.FC<PlansListProps> = ({ onError, onPlanSelection, onLogin
   }
 
   /**
-   * Callback triggered when the user select a group to filter the current list
+   * Callback triggered when the user selects a group to filter the current list
    */
   const handleFilterByGroup = (groupId: number): void => {
     setGroupFilter(groupId);
+  }
+
+  /**
+   * Callback triggered when the user selects a duration to filter the current list
+   */
+  const handleFilterByDuration = (plansIds: Array<number>): void => {
+    setPlansFilter(plansIds);
+  }
+
+  /**
+   * Callback for filtering plans to display, depending on the filter-by-plans-ids selection
+   * @see https://developer.mozilla.org/fr/docs/Web/JavaScript/Reference/Global_Objects/Array/filter
+   */
+  const filterPlan = (plan: Plan): boolean => {
+    if (!plansFilter) return true;
+
+    return plansFilter.includes(plan.id);
   }
 
   /**
@@ -193,7 +212,7 @@ const PlansList: React.FC<PlansListProps> = ({ onError, onPlanSelection, onLogin
         {categoryPlans.length === 0 && <span className="no-plans">
           {t('app.public.plans.no_plans')}
         </span>}
-        {categoryPlans.sort(comparePlans).map(plan => (
+        {categoryPlans.filter(filterPlan).sort(comparePlans).map(plan => (
           <PlanCard key={plan.id}
                     userId={customer?.id}
                     subscribedPlanId={subscribedPlanId}
@@ -209,7 +228,7 @@ const PlansList: React.FC<PlansListProps> = ({ onError, onPlanSelection, onLogin
 
   return (
     <div className="plans-list">
-      {groups && <PlansFilter user={customer} groups={groups} onGroupSelected={handleFilterByGroup} />}
+      {groups && <PlansFilter user={customer} groups={groups} onGroupSelected={handleFilterByGroup} onError={onError} onDurationSelected={handleFilterByDuration} />}
       {plans && Array.from(filteredPlans()).map(([groupId, plansByGroup]) => {
         return (
           <div key={groupId} className="plans-per-group">